+++ data/exposed-r2dbc-redisson/src/main/kotlin/io/bluetape4k/exposed/r2dbc/redisson/map/SuspendedEntityMapLoaders.kt
@@ -1,0 +1,4 @@
+package io.bluetape4k.exposed.r2dbc.redisson.map
+
+class SuspendedEntityMapLoaders {
+}
